backend/requirements.txt
asgiref==3.7.2
asttokens==2.4.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==41.0.7
decorator==5.1.1
defusedxml==0.7.1
Django==4.2.7
django-allauth==0.54.0
django-cors-headers==4.3.1
django-extensions==3.2.3
django-rest-auth==0.9.5
django-rest-auth-forked==0.9.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
executing==2.0.1
idna==3.6
ipython==8.18.1
jedi==0.19.1
Markdown==3.5.1
matplotlib-inline==0.1.6
oauthlib==3.2.2
parso==0.8.3
Pillow==10.1.0
prompt-toolkit==3.0.41
pure-eval==0.2.2
pycparser==2.21
Pygments==2.17.2
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
traitlets==5.14.0
tzdata==2023.3
urllib3==2.1.0
wcwidth==0.2.12
whitenoise==6.6.0
